break up file a bit more
