+++ src/components/organizations/LoadedOrganization.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import { Organization } from './organization';
+import { headerContainerChildStyle, statementStyle, urlStyle, imageStyle } from '../search/HPSearchResult';
+import { viewOrgStatementStyle } from "./viewOrgStatementStyle";
+export const LoadedOrganization = ({ creatorEmail, title, statement, imageThumbnailURL, organizationURL }: Organization) => <div>
+    <div style={{ display: "inline" }}>
+        <img style={imageStyle} src={imageThumbnailURL} />
+        <h1 style={headerContainerChildStyle}>{title}</h1>
+    </div>
+    <p style={viewOrgStatementStyle}>{statement}</p>
+    <a style={urlStyle} href={organizationURL}>{organizationURL}</a>
+    <p style={statementStyle}>{creatorEmail}</p>
+</div>;
