+++ src/components/organizations/MissingOrganization.tsx
@@ -1,0 +1,3 @@
+import * as React from 'react';
+
+export const MissingOrganization = () => <p>Organization Not Found!</p>
