+++ src/components/organizations/findOrgByHref.ts
@@ -1,0 +1,6 @@
+import { List } from '../../utils/list';
+import { Organization } from './organization';
+
+export function findOrgByHref(organizations: Organization[], href: string) {
+    return List(organizations).first(o => o.href == href);
+}
