+++ src/components/organizations/viewOrgStatementStyle.ts
@@ -1,0 +1,6 @@
+import { statementStyle } from '../search/HPSearchResult';
+
+export const viewOrgStatementStyle = {
+    ...statementStyle,
+    "width": "400px"
+};
